Move V8SecurityFeature into V8 directory
It is only used for arango_v8 and therefore belongs here
