+++ lib/V8/V8SecurityFeature.cpp
@@ -31,7 +31,7 @@
 #include <stdexcept>
 #include <utility>
 
-#include "ApplicationFeatures/V8SecurityFeature.h"
+#include "V8/V8SecurityFeature.h"
 
 #include "ApplicationFeatures/ApplicationServer.h"
 #include "Basics/FileUtils.h"
