+++ lib/V8/v8-environment.cpp
@@ -29,7 +29,7 @@
 #include <v8.h>
 
 #include "ApplicationFeatures/ApplicationServer.h"
-#include "ApplicationFeatures/V8SecurityFeature.h"
+#include "V8/V8SecurityFeature.h"
 #include "Basics/Common.h"
 #include "Basics/debugging.h"
 #include "V8/v8-globals.h"
