+++ lib/V8/v8-utils.cpp
@@ -62,7 +62,7 @@
 
 #include "ApplicationFeatures/ApplicationServer.h"
 #include "ApplicationFeatures/HttpEndpointProvider.h"
-#include "ApplicationFeatures/V8SecurityFeature.h"
+#include "V8/V8SecurityFeature.h"
 #include "Basics/Exceptions.h"
 #include "Basics/FileResultString.h"
 #include "Basics/FileUtils.h"
